Parse out approve/unapprove witness operations, remove margin on UserActivityActions, add % upvotes/downvotes with tooltip, refactor out some messages to components
