+++ src/activity/UserActivityActions.less
@@ -5,7 +5,6 @@
   background-color: @white;
   border: 1px solid @white-gainsboro;
   border-radius: @border-radius-base;
-  margin: 20px 0;
 
   &:empty {
     display: none;
