+++ src/constants/accountHistory.js
@@ -8,6 +8,7 @@ export const FOLLOW = 'follow';
 export const REBLOG = 'reblog';
 export const CURATION_REWARD = 'curation_reward';
 export const AUTHOR_REWARD = 'author_reward';
+export const ACCOUNT_WITNESS_VOTE = 'account_witness_vote';
 
 // Wallet Action Types
 export const TRANSFER = 'transfer';
@@ -33,4 +34,5 @@ export const PARSED_PROPERTIES = [
   TRANSFER_TO_SAVINGS,
   DELEGATE_VESTING_SHARES,
   CLAIM_REWARD_BALANCE,
+  ACCOUNT_WITNESS_VOTE,
 ];
